requirements.txt
bibtexparser
dataclasses; python_version < '3.7'
langid
openpyxl
tabulate
PyYAML
